Move tfl_quantization_driver, originally from quantization_lib, to the rest of the quantization_lib fork in lite
PiperOrigin-RevId: 735960370
