+++ tensorflow/compiler/mlir/lite/quantization/common/quantization_lib/tfl_quantization_driver.h
@@ -13,8 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef TENSORFLOW_COMPILER_MLIR_LITE_TRANSFORMS_TFL_QUANTIZATION_DRIVER_H_
-#define TENSORFLOW_COMPILER_MLIR_LITE_TRANSFORMS_TFL_QUANTIZATION_DRIVER_H_
+#ifndef TENSORFLOW_COMPILER_MLIR_LITE_QUANTIZATION_COMMON_QUANTIZATION_LIB_TFL_QUANTIZATION_DRIVER_H_
+#define TENSORFLOW_COMPILER_MLIR_LITE_QUANTIZATION_COMMON_QUANTIZATION_LIB_TFL_QUANTIZATION_DRIVER_H_
 
 #include <memory>
 #include <utility>
@@ -383,4 +383,4 @@ void ApplyQuantizationParamsPropagation(
 }  // namespace TFL
 }  // namespace mlir
 
-#endif  // TENSORFLOW_COMPILER_MLIR_LITE_TRANSFORMS_TFL_QUANTIZATION_DRIVER_H_
+#endif  // TENSORFLOW_COMPILER_MLIR_LITE_QUANTIZATION_COMMON_QUANTIZATION_LIB_TFL_QUANTIZATION_DRIVER_H_
