+++ tensorflow/compiler/mlir/lite/quantization/lite/toco_legacy/BUILD
@@ -15,6 +15,7 @@ cc_library(
     srcs = ["portable_tensor_utils.cc"],
     hdrs = ["portable_tensor_utils.h"],
     visibility = [
+        "//tensorflow/compiler/mlir/lite/quantization/common/quantization_lib:__pkg__",
         "//tensorflow/compiler/mlir/quantization/common/quantization_lib:__pkg__",
     ],
 )
